+++ src/main/kotlin/dev/codeswamp/global/auth/domain/model/authToken/RawRefreshToken.kt
@@ -1,6 +1,5 @@
 package dev.codeswamp.global.auth.domain.model.authToken
 
-import dev.codeswamp.global.auth.domain.model.AuthUser
 import java.time.Instant
 
 data class RawRefreshToken (
